requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.4.2
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
tiktoken==0.5.2 